app/src/test/java/com/example/softwaresolutionssquad/KeywordFilterControllerTest.java
package com.example.softwaresolutionssquad;

import org.junit.Before;
import org.junit.Test;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;

import static org.mockito.Mockito.*;
import static org.junit.Assert.*;

import android.content.Context;
import android.text.Editable;
import android.view.View;
import android.widget.EditText;
import android.widget.LinearLayout;
import android.widget.ListView;
import android.widget.TextView;

import com.example.softwaresolutionssquad.controllers.KeywordFilterController;
import com.example.softwaresolutionssquad.models.InventoryItem;
import com.example.softwaresolutionssquad.views.InventoryListAdapter;

import java.util.ArrayList;

public class KeywordFilterControllerTest {

    @Mock private Context mockContext;
    @Mock private LinearLayout mockKeyFilter;
    @Mock private EditText mockKeywords;
    @Mock private TextView mockKeywordButton;
    @Mock private TextView mockDateButton;
    @Mock private TextView mockMakeButton;
    @Mock private TextView mockTagButton;
    @Mock private LinearLayout mockDateFilter;
    @Mock private LinearLayout mockMakeFilter;
    @Mock private LinearLayout mockTagFilter;
    @Mock private InventoryListAdapter mockInventoryListAdapter;
    @Mock private ListView mockInventoryListView;
    private ArrayList<InventoryItem> inventoryItems;
    private KeywordFilterController controller;

    @Before
    public void setUp() {
        MockitoAnnotations.initMocks(this);

        inventoryItems = new ArrayList<>();
        // Optionally add mock inventory items to the list
        Editable editable = mock(Editable.class);
        when(editable.toString()).thenReturn("some keyword");
        when(mockKeywords.getText()).thenReturn(editable);

        controller = new KeywordFilterController(
                mockContext,
                mockKeyFilter,
                mockKeywords,
                mockKeywordButton,
                mockDateButton,
                mockMakeButton,
                mockTagButton,
                mockDateFilter,
                mockMakeFilter,
                mockTagFilter,
                mockInventoryListAdapter,
                mockInventoryListView,
                inventoryItems
        );

    }

    @Test
    public void testConstructorInitialization() {
        assertNotNull(controller);
        // Add more assertions as needed to verify initial state
    }

    @Test
    public void testToggleFilters() {
        // Initial state should be hidden
        assertEquals(0, mockKeyFilter.getVisibility());

        // Simulate button click to toggle filter visibility
        controller.toggleFilters();
        assertEquals(View.VISIBLE, mockKeyFilter.getVisibility());

        // Simulate button click to toggle filter visibility back
        controller.toggleFilters();
        assertEquals(0, mockKeyFilter.getVisibility());
    }

    @Test
    public void testApplyKeywordFilter() {
        // Set up a keyword
        String keyword = "test";
        when(mockKeywords.getText().toString()).thenReturn(keyword);

        // Trigger the filter
        controller.applyKeywordFilter(keyword);
        // Verify that the inventory list is updated
        // This assumes that your InventoryListAdapter correctly handles the filtered list
        verify(mockInventoryListView).setAdapter(any(InventoryListAdapter.class));
    }

    // More test methods can be added as needed

}
